Refactor: Replace EditorHeader with MainHeader component in PromptEditorLayout
