+++ src/components/layout/PromptEditorLayout.vue
@@ -2,7 +2,7 @@
   <div
     class="bg-gray-900 text-gray-200 font-sans antialiased min-h-screen flex flex-col"
   >
-    <EditorHeader />
+    <MainHeader />
     <main class="flex-grow flex flex-col md:flex-row gap-4 p-4 overflow-hidden">
       <slot name="prompt-text" />
       <slot name="tag-block" />
@@ -11,5 +11,5 @@
 </template>
 
 <script setup lang="ts">
-import EditorHeader from './EditorHeader.vue'
+import MainHeader from './MainHeader.vue'
 </script>
